+++ src/main/java/toni/doesittick/api/Tickable.java
@@ -2,10 +2,8 @@ package toni.doesittick.api;
 
 public final class Tickable {
     public interface EntityType {
-        boolean doespotatotick$shouldAlwaysTick();
-        void doespotatotick$setShouldAlwaysTick();
-        boolean doespotatotick$shouldAlwaysTickInRaid();
-        void doespotatotick$setShouldAlwaysTickInRaid();
+        Boolean doespotatotick$shouldAlwaysTick();
+        Boolean doespotatotick$shouldAlwaysTickInRaid();
     }
 
     public interface Level {