+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-## Changelog
-- Initial release
+1.1.4 - Disable on client by default
+1.1.3 - Increase default ticking distance around chunks marked as active
+1.1.2 - Fix an issue with falling block entities building up due to not ticking
+1.1.1 - Fixed config whitelist not loading on Forge
+1.1.0 - Added support & dependency for Chunk Activity Tracker, which will turn off Does It Tick automatically in frequently visited chunks.